requirements.txt
disnake~=2.9.2
openai~=1.30.2
python-dotenv~=1.0.1